win32: move WIN10_BUILD_VERSION to `util.rs` for reusability
